requirements.txt
streamlit
ultralytics
Pillow
torch
gspread
google-api-python-client
google-auth